feat(celery): rename one celery app to then add operational info deletion
