+++ entrypoint.sh
@@ -16,7 +16,7 @@ case "$1" in
             --workers=${API_WORKERS} \
             --worker-class=uvicorn.workers.UvicornWorker ;;
     beat) poetry run celery beat \
-            --app=immuni_analytics.celery.exposure_payload.app.celery_app \
+            --app=immuni_analytics.celery.scheduled.app.celery_app \
             --loglevel=debug ;;
     worker) poetry run celery worker \
             --app=immuni_analytics.celery.${CELERY_APP_NAME}.app.celery_app \
