+++ immuni_analytics/core/config.py
@@ -41,8 +41,8 @@ CELERY_ALWAYS_EAGER: bool = config(
 CELERY_BROKER_REDIS_URL_AUTHORIZATION: str = config(
     "CELERY_BROKER_REDIS_URL_AUTHORIZATION", default="redis://localhost:6379/0"
 )
-CELERY_BROKER_REDIS_URL_EXPOSURE_PAYLOAD: str = config(
-    "CELERY_BROKER_REDIS_URL_EXPOSURE_PAYLOAD", default="redis://localhost:6379/0"
+CELERY_BROKER_REDIS_URL_SCHEDULED: str = config(
+    "CELERY_BROKER_REDIS_URL_SCHEDULED", default="redis://localhost:6379/0"
 )
 CELERY_BROKER_REDIS_URL_OPERATIONAL_INFO: str = config(
     "CELERY_BROKER_REDIS_URL_OPERATIONAL_INFO", default="redis://localhost:6379/0"
