+++ tests/conftest.py
@@ -28,7 +28,7 @@ check_redis_url(
     config.CELERY_BROKER_REDIS_URL_AUTHORIZATION, "CELERY_BROKER_REDIS_URL_AUTHORIZATION"
 )
 check_redis_url(
-    config.CELERY_BROKER_REDIS_URL_EXPOSURE_PAYLOAD, "CELERY_BROKER_REDIS_URL_EXPOSURE_PAYLOAD"
+    config.CELERY_BROKER_REDIS_URL_SCHEDULED, "CELERY_BROKER_REDIS_URL_SCHEDULED"
 )
 check_redis_url(
     config.CELERY_BROKER_REDIS_URL_OPERATIONAL_INFO, "CELERY_BROKER_REDIS_URL_OPERATIONAL_INFO"
