+++ tests/fixtures/core.py
@@ -55,7 +55,7 @@ def client(
 
 @fixture(scope="function")
 def setup_exposure_payload_celery_app(monitoring_setup: None) -> Celery:
-    from immuni_analytics.celery.exposure_payload.app import celery_app
+    from immuni_analytics.celery.scheduled.app import celery_app
 
     celery_app.conf.update(CELERY_ALWAYS_EAGER=True)
     return celery_app
